feat: support new and old architectures on iOS
