+++ ios/Emm-Bridging-Header.h
@@ -1,2 +1,1 @@
-#import <React/RCTBridgeModule.h>
-#import <React/RCTEventEmitter.h>
+
