+++ package.json
@@ -33,7 +33,7 @@
     "prepare": "bob build",
     "release": "release-it",
     "example": "cd example && npm",
-    "pods": "cd example && pod-install --quiet",
+    "pods": "cd example && RCT_NEW_ARCH_ENABLED=1 pod-install --quiet",
     "bootstrap": "npm run example i && npm run pods"
   },
   "keywords": [
@@ -50,32 +50,37 @@
   },
   "homepage": "https://github.com/mattermost/react-native-emm#readme",
   "devDependencies": {
-    "@commitlint/config-conventional": "19.1.0",
-    "@react-native/eslint-config": "0.73.2",
-    "@react-native/metro-config": "0.73.5",
+    "@commitlint/config-conventional": "19.2.2",
+    "@react-native/eslint-config": "0.74.83",
+    "@react-native/metro-config": "0.74.83",
     "@release-it/conventional-changelog": "8.0.1",
     "@types/jest": "29.5.12",
-    "@types/react": "18.2.75",
-    "@typescript-eslint/eslint-plugin": "6.21.0",
-    "@typescript-eslint/parser": "6.21.0",
-    "commitlint": "19.2.1",
+    "@types/react": "18.3.1",
+    "@typescript-eslint/eslint-plugin": "7.8.0",
+    "@typescript-eslint/parser": "7.8.0",
+    "commitlint": "19.3.0",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-flowtype": "8.0.3",
-    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-prettier": "5.1.3",
     "husky": "9.0.11",
     "jest": "29.7.0",
-    "pod-install": "0.2.0",
-    "prettier": "2.8.8",
+    "pod-install": "0.2.2",
+    "prettier": "3.2.5",
     "react": "18.2.0",
-    "react-native": "0.73.6",
+    "react-native": "0.74.1",
     "react-native-builder-bob": "0.23.2",
-    "release-it": "17.1.1",
-    "typescript": "5.3.3"
+    "release-it": "17.2.1",
+    "typescript": "5.4.5"
   },
   "peerDependencies": {
     "react": "*",
     "react-native": "*"
+  },
+  "codegenConfig": {
+    "name": "EmmSpec",
+    "type": "modules",
+    "jsSrcsDir": "./src"
   },
   "jest": {
     "preset": "react-native",
