+++ react-native-emm.podspec
@@ -17,6 +17,23 @@ Pod::Spec.new do |s|
   
   s.source_files = "ios/**/*.{h,m,mm,swift}"
   
+  fabric_enabled = ENV["RCT_NEW_ARCH_ENABLED"] == "1"
 
-  s.dependency "React-Core"
+  if fabric_enabled
+    s.pod_target_xcconfig    = {
+      "DEFINES_MODULE" => "YES",
+      "BUILD_LIBRARY_FOR_DISTRIBUTION" => "YES",
+      "OTHER_CPLUSPLUSFLAGS" => "-DRCT_NEW_ARCH_ENABLED=1",
+      "OTHER_SWIFT_FLAGS" => "-no-verify-emitted-module-interface"
+    }
+  else
+    s.pod_target_xcconfig    = {
+      "DEFINES_MODULE" => "YES",
+      "BUILD_LIBRARY_FOR_DISTRIBUTION" => "YES",
+      "OTHER_SWIFT_FLAGS" => "-no-verify-emitted-module-interface"
+    }
+  end
+
+  install_modules_dependencies(s)
+
 end
