+++ src/emm-native.ts
@@ -1,0 +1,1 @@
+export default require("./NativeEmm").default;
